sched/fair: fix misfit with PELT
Commit 20017f33839a ("sched/fair: Only kick nohz balance when runqueue
has more than 1 task") disabled the nohz kick for LB when a rq has a
misfit task. The assumption is that this would be addressed in the
forced up-migration path. However, this path is WALT-specific, so
disabling the nohz kick breaks PELT.

Fix it by re-enabling the nohz_kick when there is a misfit task on the
rq.

Bug: 143472450
Bug: 145190765
Test: 10/10 iterations of eas_small_to_big ended up up-migrating
Fixes: 20017f33839a ("sched/fair: Only kick nohz balance when runqueue
has more than 1 task")
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: I9f708eb7661a9e82afdd4e99b878995c33703a45
(cherry picked from commit 2639b055d2526603ec8c12a2fe2413bf445c3def)
Signed-off-by: Kyle Lin <kylelin@google.com>
Signed-off-by: Excalibur-99 <txexcalibur99@gmail.com>